refactor: new client package
