+++ internal/pkg/client/http/client.go
@@ -1,0 +1,71 @@
+// ------------------------------------------------------------------------
+// SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
+// SPDX-FileName: internal/pkg/client/http/fetch.go
+// SPDX-FileType: SOURCE
+// SPDX-License-Identifier: Apache-2.0
+// ------------------------------------------------------------------------
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ------------------------------------------------------------------------
+package http
+
+import (
+	"fmt"
+	"regexp"
+
+	"github.com/bomctl/bomctl/internal/pkg/url"
+)
+
+type Client struct{}
+
+func (*Client) Name() string {
+	return "HTTP"
+}
+
+func (*Client) RegExp() *regexp.Regexp {
+	return regexp.MustCompile(
+		fmt.Sprintf("%s%s%s%s",
+			`((?P<scheme>https?):\/\/)`,
+			`((?P<username>[^:]+)(?::(?P<password>[^@]+))?(?:@))?`,
+			`(?P<hostname>[^@\/?#:]+)(?::(?P<port>\d+))?`,
+			`(\/?(?P<path>[^@?#]*))(\?(?P<query>[^#]*))?(#(?P<fragment>.*))?`,
+		),
+	)
+}
+
+func (client *Client) Parse(fetchURL string) *url.ParsedURL {
+	results := map[string]string{}
+	pattern := client.RegExp()
+	match := pattern.FindStringSubmatch(fetchURL)
+
+	for idx, name := range match {
+		results[pattern.SubexpNames()[idx]] = name
+	}
+
+	// Ensure required map fields are present.
+	for _, required := range []string{"scheme", "hostname"} {
+		if value, ok := results[required]; !ok || value == "" {
+			return nil
+		}
+	}
+
+	return &url.ParsedURL{
+		Scheme:   results["scheme"],
+		Username: results["username"],
+		Password: results["password"],
+		Hostname: results["hostname"],
+		Port:     results["port"],
+		Path:     results["path"],
+		Query:    results["query"],
+		Fragment: results["fragment"],
+	}
+}
