+++ internal/pkg/client/http/fetch.go
@@ -1,6 +1,6 @@
 // ------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/fetch/http/http.go
+// SPDX-FileName: internal/pkg/client/http/fetch.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // ------------------------------------------------------------------------
@@ -23,57 +23,11 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"regexp"
 
 	"github.com/bomctl/bomctl/internal/pkg/url"
 )
 
-type Fetcher struct{}
-
-func (*Fetcher) Name() string {
-	return "HTTP"
-}
-
-func (*Fetcher) RegExp() *regexp.Regexp {
-	return regexp.MustCompile(
-		fmt.Sprintf("%s%s%s%s",
-			`((?P<scheme>https?):\/\/)`,
-			`((?P<username>[^:]+)(?::(?P<password>[^@]+))?(?:@))?`,
-			`(?P<hostname>[^@\/?#:]+)(?::(?P<port>\d+))?`,
-			`(\/?(?P<path>[^@?#]*))(\?(?P<query>[^#]*))?(#(?P<fragment>.*))?`,
-		),
-	)
-}
-
-func (fetcher *Fetcher) Parse(fetchURL string) *url.ParsedURL {
-	results := map[string]string{}
-	pattern := fetcher.RegExp()
-	match := pattern.FindStringSubmatch(fetchURL)
-
-	for idx, name := range match {
-		results[pattern.SubexpNames()[idx]] = name
-	}
-
-	// Ensure required map fields are present.
-	for _, required := range []string{"scheme", "hostname"} {
-		if value, ok := results[required]; !ok || value == "" {
-			return nil
-		}
-	}
-
-	return &url.ParsedURL{
-		Scheme:   results["scheme"],
-		Username: results["username"],
-		Password: results["password"],
-		Hostname: results["hostname"],
-		Port:     results["port"],
-		Path:     results["path"],
-		Query:    results["query"],
-		Fragment: results["fragment"],
-	}
-}
-
-func (*Fetcher) Fetch(parsedURL *url.ParsedURL, auth *url.BasicAuth) ([]byte, error) {
+func (*Client) Fetch(parsedURL *url.ParsedURL, auth *url.BasicAuth) ([]byte, error) {
 	req, err := http.NewRequestWithContext(context.Background(), "GET", parsedURL.String(), nil)
 	if err != nil {
 		return nil, fmt.Errorf("failed creating request to %s: %w", parsedURL.String(), err)
